libopenldap: Add missing Config.in hook
Package has a consistent Config.in file, but it isn't added to all parts of the file.

libopenldap is unable to be built with --enable-debug because the Config.in is never
attached to the libopenldap.

Signed-off-by: Donald Hoskins <grommish@gmail.com>
(cherry picked from commit 50a5db44b96882c39ba24fd0f324868f84a0d3e8)